requirements.txt
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cloudscraper==1.2.71
colorama==0.4.6
editdistpy==0.1.3
gevent==23.9.1
greenlet==2.0.2
idna==3.4
numpy==1.25.2
pandas==2.0.3
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
requests-toolbelt==1.0.0
requests==2.31.0
sentencepiece==0.1.99
setuptools==65.5.1
six==1.16.0
sseclient-py
tomli==2.0.1
tqdm==4.65.1
tzdata==2023.3
urllib3==2.0.4
websocket-client==1.6.3
wheel==0.38.4
zope.event==5.0
zope.interface==6.0
